Control flow graph loop pattern detection fix; Improvement of "Native" JavaScript bindings; Port of JavaScript BigInteger library (will be used as backend for decimal implementation);

+++ NetWebScript/NetWebScript.Core/Script/JSNumber.cs
@@ -83,7 +83,7 @@ namespace NetWebScript.Script
             }
             try
             {
-                return new JSNumber(Convert.ToInt64(s, radix));
+                return new JSNumber(Convert.ToInt32(s, radix));
             }
             catch
             {
